refact: Moving LinearFit into its own sub-module
